Consolidate Arbitrary impls and property tests into single module
Move 14 Arbitrary implementations and 39 quickcheck property tests from
7 scattered files into a unified property_tests.rs module. Also clean up
redundant grid tests by removing 5 covered tests and consolidating 6
related tests into 3.
