+++ crates/gdsr/src/lib.rs
@@ -6,6 +6,8 @@ mod grid;
 mod library;
 mod point;
 #[cfg(test)]
+mod property_tests;
+#[cfg(test)]
 pub(crate) mod test_fixtures;
 mod traits;
 mod transformation;
