+++ crates/gdsr/src/units.rs
@@ -284,44 +284,6 @@ mod tests {
 
     use super::*;
 
-    use quickcheck::{Arbitrary, Gen};
-
-    const MAX_VALUE: i32 = 10_000;
-
-    impl Arbitrary for IntegerUnit {
-        fn arbitrary(g: &mut Gen) -> Self {
-            let value = (i32::arbitrary(g) % MAX_VALUE).clamp(-MAX_VALUE, MAX_VALUE);
-            let units_options = [1e-9, 1e-8, 1e-7, 1e-6];
-            let units = units_options[usize::arbitrary(g) % units_options.len()];
-            Self { value, units }
-        }
-    }
-
-    impl Arbitrary for FloatUnit {
-        fn arbitrary(g: &mut Gen) -> Self {
-            let raw_value = f64::arbitrary(g);
-            let value = if raw_value.is_finite() {
-                (raw_value % f64::from(MAX_VALUE))
-                    .clamp(f64::from(-MAX_VALUE), f64::from(MAX_VALUE))
-            } else {
-                0.0 // Replace NaN/inf with 0
-            };
-            let units_options = [1e-9, 1e-8, 1e-7, 1e-6];
-            let units = units_options[usize::arbitrary(g) % units_options.len()];
-            Self { value, units }
-        }
-    }
-
-    impl Arbitrary for Unit {
-        fn arbitrary(g: &mut Gen) -> Self {
-            if bool::arbitrary(g) {
-                Self::Integer(IntegerUnit::arbitrary(g))
-            } else {
-                Self::Float(FloatUnit::arbitrary(g))
-            }
-        }
-    }
-
     mod creation {
         use super::*;
 
@@ -809,142 +771,6 @@ mod tests {
             let scaled = unit.scale_to(1e-6);
 
             assert_eq!(scaled, Unit::float(1.0, 1e-6));
-        }
-    }
-
-    mod property_tests {
-        use super::*;
-        use quickcheck_macros::quickcheck;
-
-        fn approx_eq(a: f64, b: f64, epsilon: f64) -> bool {
-            (a - b).abs() < epsilon
-        }
-
-        #[quickcheck]
-        fn float_addition_commutativity(a: FloatUnit, b: FloatUnit) -> bool {
-            let a = Unit::Float(a);
-            let b = Unit::Float(b);
-            let sum_a_b = a + b;
-            let sum_b_a = b + a;
-
-            sum_a_b == sum_b_a
-        }
-
-        #[quickcheck]
-        fn float_addition_associativity(a: FloatUnit, b: FloatUnit, c: FloatUnit) -> bool {
-            let a = Unit::Float(a);
-            let b = Unit::Float(b);
-            let c = Unit::Float(c);
-            let sum_ab_c = (a + b) + c;
-            let sum_a_bc = a + (b + c);
-
-            sum_a_bc == sum_ab_c
-        }
-
-        #[quickcheck]
-        fn addition_identity(a: Unit) -> bool {
-            let zero = match a {
-                Unit::Integer(_) => Unit::integer(0, a.units()),
-                Unit::Float(_) => Unit::float(0.0, a.units()),
-            };
-
-            let a_plus_zero = a + zero;
-            let zero_plus_a = zero + a;
-            zero_plus_a == a && a_plus_zero == zero_plus_a
-        }
-
-        #[quickcheck]
-        fn multiplicative_identity(a: Unit) -> bool {
-            let one = match a {
-                Unit::Integer(_) => Unit::integer(1, a.units()),
-                Unit::Float(_) => Unit::float(1.0, a.units()),
-            };
-
-            let a_times_one = a * one;
-            let one_times_a = one * a;
-            one_times_a == a && a_times_one == one_times_a
-        }
-
-        #[quickcheck]
-        fn addition_preserves_units(a: Unit, b: Unit) -> bool {
-            let result = a + b;
-            result.units() == a.units()
-        }
-
-        #[quickcheck]
-        fn subtraction_preserves_units(a: Unit, b: Unit) -> bool {
-            let result = a - b;
-            result.units() == a.units()
-        }
-
-        /// Verifies that the absolute value of the sum equals the sum of absolute values.
-        /// Tolerance accounts for unit scaling and integer rounding.
-        #[quickcheck]
-        fn addition_absolute_value_correctness(a: Unit, b: Unit) -> bool {
-            let sum = a + b;
-            let abs_sum = sum.absolute_value();
-            let abs_a = a.absolute_value();
-            let abs_b = b.absolute_value();
-            let expected = abs_a + abs_b;
-
-            approx_eq(abs_sum, expected, a.units())
-        }
-
-        /// Verifies that a + (-a) ≈ 0.
-        #[quickcheck]
-        fn addition_inverse_property(a: Unit) -> bool {
-            let neg_a = match a {
-                Unit::Integer(IntegerUnit { value, units }) => Unit::Integer(IntegerUnit {
-                    value: -value,
-                    units,
-                }),
-                Unit::Float(FloatUnit { value, units }) => Unit::Float(FloatUnit {
-                    value: -value,
-                    units,
-                }),
-            };
-
-            let result = a + neg_a;
-            let abs_result = result.absolute_value().abs();
-            abs_result < 1e-9
-        }
-
-        /// Verifies exact integer addition when units are identical.
-        #[quickcheck]
-        fn addition_same_units_no_scaling_error_integer(a: IntegerUnit, b: IntegerUnit) -> bool {
-            let a_unit = Unit::Integer(IntegerUnit {
-                value: a.value,
-                units: a.units,
-            });
-            let b_unit = Unit::Integer(IntegerUnit {
-                value: b.value,
-                units: a.units,
-            });
-
-            let result = a_unit + b_unit;
-
-            let integer_result = result.to_integer_unit();
-
-            integer_result == Unit::integer(a.value + b.value, a.units)
-        }
-
-        /// Verifies precise float addition when units are identical.
-        #[quickcheck]
-        fn addition_same_units_no_scaling_error_float(a: FloatUnit, b: FloatUnit) -> bool {
-            let a_unit = Unit::Float(FloatUnit {
-                value: a.value,
-                units: a.units,
-            });
-            let b_unit = Unit::Float(FloatUnit {
-                value: b.value,
-                units: a.units,
-            });
-
-            let result = a_unit + b_unit;
-
-            let float_result = result.to_float_unit();
-
-            float_result == Unit::float(a.value + b.value, a.units)
         }
     }
 }
